requirements.txt
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
tiktoken>=0.7.0
